# AP_DDS: publish gps global origin