@srmainwaring 2024-03-06
> 1. Can you update the README with the new topic

Updated in https://github.com/ArduPilot/ardupilot/pull/26409/commits/3d0d9a7c5188fd4650617b87002896b5885f2f5d